Packages: Rename spec-parser to better explain its purpose
