+++ blocks/api/test/parser.js
@@ -547,7 +547,7 @@ describe( 'block parser', () => {
 		} );
 	} );
 
-	describe( 'parse() of @wordpress/spec-parser', () => {
+	describe( 'parse() of @wordpress/block-serialization-spec-parser', () => {
 		// run the test cases using the PegJS defined parser
 		testCases( parsePegjs );
 	} );
