+++ package.json
@@ -19,6 +19,7 @@
 		"@wordpress/api-fetch": "file:packages/api-fetch",
 		"@wordpress/autop": "file:packages/autop",
 		"@wordpress/blob": "file:packages/blob",
+		"@wordpress/block-serialization-spec-parser": "file:packages/block-serialization-spec-parser",
 		"@wordpress/components": "file:packages/components",
 		"@wordpress/compose": "file:packages/compose",
 		"@wordpress/core-data": "file:packages/core-data",
@@ -35,7 +36,6 @@
 		"@wordpress/keycodes": "file:packages/keycodes",
 		"@wordpress/plugins": "file:packages/plugins",
 		"@wordpress/shortcode": "file:packages/shortcode",
-		"@wordpress/spec-parser": "file:packages/spec-parser",
 		"@wordpress/url": "file:packages/url",
 		"@wordpress/viewport": "file:packages/viewport",
 		"@wordpress/wordcount": "file:packages/wordcount",
