+++ packages/block-serialization-spec-parser/README.md
@@ -1,6 +1,6 @@
-# @wordpress/spec-parser
+# @wordpress/block-serialization-spec-parser
 
-This library contains the grammar file (`grammar.pegjs`) for WordPress posts which is a _specification_ which is used to generate the actual _parser_ which is also bundled in this package.
+This library contains the grammar file (`grammar.pegjs`) for WordPress posts which is a block serialization _specification_ which is used to generate the actual _parser_ which is also bundled in this package.
 
 PEG parser generators are available in many languages, though different libraries may require some translation of this grammar into their syntax. For more information see:
 * https://pegjs.org
@@ -11,13 +11,13 @@ PEG parser generators are available in many languages, though different librarie
 Install the module
 
 ```bash
-npm install @wordpress/spec-parser --save
+npm install @wordpress/block-serialization-spec-parser --save
 ```
 
 ## Usage
 
 ```js
-import { parse } from '@wordpress/spec-parser';
+import { parse } from '@wordpress/block-serialization-spec-parser';
 
 parse( '<!-- wp:core/more --><!--more--><!-- /wp:core/more -->' );
 // [{"attrs": null, "blockName": "core/more", "innerBlocks": [], "innerHTML": "<!--more-->"}]
