+++ packages/block-serialization-spec-parser/index.js
