+++ packages/block-serialization-spec-parser/package.json
@@ -1,14 +1,16 @@
 {
-	"name": "@wordpress/spec-parser",
+	"name": "@wordpress/block-serialization-spec-parser",
 	"version": "1.0.0-alpha.1",
-	"description": "Parser for WordPress posts",
+	"description": "Block serialization specification parser for WordPress posts",
 	"author": "The WordPress Contributors",
 	"license": "GPL-2.0-or-later",
 	"keywords": [
 		"wordpress",
+		"block",
+		"spec",
 		"parser"
 	],
-	"homepage": "https://github.com/WordPress/gutenberg/tree/master/packages/spec-parser/README.md",
+	"homepage": "https://github.com/WordPress/gutenberg/tree/master/packages/block-serialization-spec-parser/README.md",
 	"repository": {
 		"type": "git",
 		"url": "https://github.com/WordPress/gutenberg.git"
